Simplify API by removing RNG interface (#8)
To improve usability of API, the existing BackoffAlgorithm_RNG_t interface for passing a random number generator function to the library is removed in this PR. This helps avoid the need of "wrappers" around platform-specific RNGs to pass to the library. Instead, the application can generate the random number using their platform-specific RNG and just directly pass to the library for calculating the next retry backoff period.