Add stdint.readme for C90 compliance (#13)
Add a stdint.readme file to support toolchains that do not provide the stdint.h file for integer typedefs. With the provision of this file, the library can be used against C90 compilers. This PR also adds a CI check to ensure successful library builds when using the stdint.readme file for the standard C library header.